Update README, add Changes.md, rename JS macro handler
